Move Eolymp Basecamp parsing into its own file
