+++ src/parsers/contest/EolympContestParser.ts
@@ -1,9 +1,9 @@
-import { EolympProblemParser } from '../problem/EolympProblemParser';
+import { EolympNormalProblemParser } from '../problem/EolympNormalProblemParser';
 import { SimpleContestParser } from '../SimpleContestParser';
 
 export class EolympContestParser extends SimpleContestParser {
   protected linkSelector = '.eo-problem-row__name';
-  protected problemParser = new EolympProblemParser();
+  protected problemParser = new EolympNormalProblemParser();
 
   public getMatchPatterns(): string[] {
     return ['https://www.eolymp.com/*/contests/*'];
